@@ -1,22 +1,22 @@
 module github.com/devopsext/events
 
-go 1.19
+go 1.21
 
 require (
-	cloud.google.com/go/pubsub v1.21.1
+	cloud.google.com/go/pubsub v1.33.0
+	github.com/IBM/sarama v1.41.2
 	github.com/Masterminds/sprig/v3 v3.2.2 // indirect
-	github.com/Shopify/sarama v1.31.1
-	github.com/VictoriaMetrics/metricsql v0.40.0
+	github.com/VictoriaMetrics/metricsql v0.65.0
 	github.com/devopsext/sre v0.1.6
-	github.com/devopsext/tools v0.8.9
-	github.com/devopsext/utils v0.2.0
-	github.com/go-playground/webhooks/v6 v6.0.0-rc.1
-	github.com/grafana-tools/sdk v0.0.0-20210521150820-354cd37a4b4e
-	github.com/prometheus/alertmanager v0.22.0
-	github.com/spf13/cobra v1.4.0
-	github.com/xanzy/go-gitlab v0.60.0
-	google.golang.org/api v0.76.0
-	google.golang.org/genproto v0.0.0-20220426171045-31bebdecfb46 // indirect
+	github.com/devopsext/tools v0.9.13
+	github.com/devopsext/utils v0.2.1
+	github.com/go-playground/webhooks/v6 v6.3.0
+	github.com/grafana-tools/sdk v0.0.0-20220919052116-6562121319fc
+	github.com/prometheus/alertmanager v0.26.0
+	github.com/spf13/cobra v1.7.0
+	github.com/xanzy/go-gitlab v0.91.1
+	google.golang.org/api v0.142.0
+	google.golang.org/genproto v0.0.0-20230920204549-e6e6cdab5c13 // indirect
 	k8s.io/api v0.23.3
 	k8s.io/apimachinery v0.23.3
 )
